# synkdata/docs/api.yaml
openapi: 3.0.3
info:
  title: SynkData Identity Intelligence Platform API
  description: |
    Plataforma de inteligencia de identidad para verificación, antecedentes,
    cumplimiento normativo y análisis de riesgo enfocada en México y LATAM.

    ## Autenticación
    Todos los endpoints protegidos requieren un token JWT en el header
    `Authorization: Bearer <token>`.

    ## Códigos de respuesta
    - `200` Operación exitosa
    - `201` Recurso creado
    - `400` Error de validación en los datos de entrada
    - `401` Token de autenticación inválido o expirado
    - `403` Permisos insuficientes para la operación
    - `404` Recurso no encontrado
    - `429` Límite de peticiones excedido
    - `500` Error interno del servidor
  version: 1.0.0
  contact:
    name: SynkData Team
    email: dev@synkdata.io
    url: https://synkdata.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Servidor de desarrollo
  - url: https://api.synkdata.io/api/v1
    description: Servidor de producción
security:
  - BearerAuth: []
...
paths:
...
  /curp/validate:
    post:
      tags: [Verification]
      summary: Validar una CURP
      description: |
        Realiza la validación completa de una CURP: formato, dígito verificador
        y opcionalmente consulta contra la base de datos de RENAPO.
      operationId: validateCurp
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurpValidateRequest'
            example:
              curp: "GOME850101HDFRRN09"
              check_renapo: false
      responses:
        '200':
          description: Resultado de la validación de la CURP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurpValidationResult'
  /curp/search:
    post:
      tags: [Verification]
      summary: Buscar CURP en RENAPO
      description: |
        Busca una CURP en la base de datos de RENAPO proporcionando
        los datos personales del individuo.
      operationId: searchCurp
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurpSearchRequest'
            example:
              name: "JUAN"
              paternal: "GOMEZ"
              maternal: "RODRIGUEZ"
              birth_date: "1985-01-01"
              gender: "H"
              state: "DF"
      responses:
        '200':
          description: Lista de candidatos CURP encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CurpCandidate'
  /curp/{curp}:
    get:
      tags: [Verification]
      summary: Obtener validación cacheada de CURP
      operationId: getCachedCurpValidation
      security:
        - BearerAuth: []
      parameters:
        - name: curp
          in: path
          required: true
          schema:
            type: string
            minLength: 18
            maxLength: 18
          description: CURP a buscar en el caché
          example: "GOME850101HDFRRN09"
      responses:
        '200':
          description: Resultado de validación cacheado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurpValidationResult'
        '404':
          description: No se encontró una validación previa
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint de autenticación.
        Formato: `Authorization: Bearer <token>`
  schemas:
...
    CurpValidateRequest:
      type: object
      required: [curp]
      properties:
        curp:
          type: string
          minLength: 18
          maxLength: 18
          description: CURP a validar
          example: "GOME850101HDFRRN09"
        check_renapo:
          type: boolean
          default: false
          description: Consultar RENAPO para validación adicional
    CurpSearchRequest:
      type: object
      required: [name, paternal, maternal, birth_date, gender]
      properties:
        name:
          type: string
          description: Nombre(s) de pila
          example: "JUAN"
        paternal:
          type: string
          description: Apellido paterno
          example: "GOMEZ"
        maternal:
          type: string
          description: Apellido materno
          example: "RODRIGUEZ"
        birth_date:
          type: string
          format: date
          description: Fecha de nacimiento
          example: "1985-01-01"
        gender:
          type: string
          pattern: "^[HM]$"
          description: "Sexo: 'H' o 'M'"
          example: "H"
        state:
          type: string
          maxLength: 2
          description: Clave de la entidad federativa (2 letras)
          example: "DF"
...
    CurpValidationResult:
      type: object
      properties:
        is_valid:
          type: boolean
          description: Si la CURP pasó todas las validaciones
        format_valid:
          type: boolean
          description: Si el formato de la CURP es correcto
        check_digit_valid:
          type: boolean
          description: Si el dígito verificador es correcto
        renapo_match:
          type: boolean
          nullable: true
          description: Si la CURP fue encontrada en RENAPO
        extracted_info:
          type: object
          nullable: true
          description: Información extraída de la CURP
          properties:
            name_initials:
              type: string
              example: "GOME"
            birth_date:
              type: string
              example: "1985-01-01"
            gender:
              type: string
              example: "H"
            state_code:
              type: string
              example: "DF"
            state_name:
              type: string
              example: "Ciudad de México"
            internal_consonants:
              type: string
              example: "RRN"
        errors:
          type: array
          items:
            type: string
          description: Lista de errores encontrados
        validated_at:
          type: string
          format: date-time
...
    CurpCandidate:
      type: object
      properties:
        curp:
          type: string
        name:
          type: string
          nullable: true
        birth_date:
          type: string
          nullable: true
        match_score:
          type: number
          minimum: 0.0
          maximum: 1.0
...